Everything that was called 'check' has been renamed to 'receipt'

+++ checksystem-core/src/main/java/ru/clevertec/checksystem/core/common/template/IPdfTemplate.java
@@ -1,7 +1,7 @@
 package ru.clevertec.checksystem.core.common.template;
 
 public interface IPdfTemplate extends ITemplate {
-    int getTopOffset();
+    long getTopOffset();
 
-    void setTopOffset(int offset);
+    void setTopOffset(long offset);
 }
